Move modal to shared component to be reused in user creation/update
