+++ admin/src/main/ts/src/app/_shared/user-position-modal/user-position-modal.component.scss
