+++ admin/src/main/ts/src/app/_shared/user-position-modal/user-position-modal.component.ts
@@ -1,10 +1,8 @@
 import {
-  AfterViewInit,
   Component,
   ElementRef,
   Inject,
   Injector,
-  Input,
   ViewChild,
 } from "@angular/core";
 import { OdeComponent } from "ngx-ode-core";
